Fixed [Datalogger]: every time that a .csv reaches 50k lines, start a new file. Add '-N' to the filename (0, 1, 2, ...)
If the file was open(in readonly) while logging, the QtextStream was stopping writing data in the file untile you closed and open another file. But he was accumulating data into his the QtextStream buffer. So, when you open another file, you had the previously accumulated data written in the new file.

Fixed.